+++ PotreeConverter/src/stuff.cpp
@@ -95,9 +95,9 @@ AABB childAABB(const AABB &aabb, const int &index){
  *
  */
 int nodeIndex(const AABB &aabb, const Point &point){
-	int mx = 2.0 * (point.x - aabb.min.x) / aabb.size.x;
-	int my = 2.0 * (point.y - aabb.min.y) / aabb.size.y;
-	int mz = 2.0 * (point.z - aabb.min.z) / aabb.size.z;
+	int mx = (int)(2.0 * (point.x - aabb.min.x) / aabb.size.x);
+	int my = (int)(2.0 * (point.y - aabb.min.y) / aabb.size.y);
+	int mz = (int)(2.0 * (point.z - aabb.min.z) / aabb.size.z);
 
 	mx = min(mx, 1);
 	my = min(my, 1);
